+++ src/components/topnav.jsx
@@ -127,6 +127,11 @@ function topnav() {
                   <i className="fa-solid fa-jet-fighter-up fa-2xl"></i>
                 </span>
               </a>
+              <a href="/sales_report" target='_blanck' className="notification" style={{ position: "relative", display: "inline-block" }}>
+                <span className="page-title">
+                <i class="fa-solid fa-trophy fa-2xl"></i>
+                </span>
+              </a>
               <a href="/live_tickets" target='_blanck' className="notification" style={{ position: "relative", display: "inline-block" }}>
                 <i className="fa-solid fa-ticket fa-2xl pointer"></i>
                 <span className='bg-danger text-white rounded-circle text-center' style={{
